+++ app/src/main/kotlin/jp/co/yumemi/android/code_check/app/GithubSampleApplication.kt
@@ -1,4 +1,4 @@
-package jp.co.yumemi.android.code_check
+package jp.co.yumemi.android.code_check.app
 
 import android.app.Application
 import dagger.hilt.android.HiltAndroidApp
