+++ app/src/main/kotlin/jp/co/yumemi/android/code_check/app/MainActivity.kt
@@ -1,10 +1,11 @@
 /*
  * Copyright © 2021 YUMEMI Inc. All rights reserved.
  */
-package jp.co.yumemi.android.code_check
+package jp.co.yumemi.android.code_check.app
 
 import androidx.appcompat.app.AppCompatActivity
 import dagger.hilt.android.AndroidEntryPoint
+import jp.co.yumemi.android.code_check.R
 
 @AndroidEntryPoint
 class MainActivity : AppCompatActivity(R.layout.activity_main)
